sdk: Add constructor from m.room.message event for Message
